+++ examples/embed-starter-kit/site/blocks/page/dynamic-columns/dynamic-columns.builder.ts
@@ -28,7 +28,7 @@ Builder.registerComponent(
           {
             name: 'image',
             type: 'file',
-            allowedFileTypes: ['jpeg', 'jpg', 'png', 'svg'],
+            allowedFileTypes: ['jpeg', 'jpg', 'png', 'svg', 'webp'],
             required: true,
             defaultValue:
               'https://cdn.builder.io/api/v1/image/assets%2Fpwgjf0RoYWbdnJSbpBAjXNRMe9F2%2Ffb27a7c790324294af8be1c35fe30f4d',