+++ examples/next-js-sdk-gen-2-experimental-app-directory/next.config.js
@@ -4,6 +4,9 @@ const nextConfig = {
    * This setting is required for Builder's Visual Editor to work with your site.
    */
   transpilePackages: ['@builder.io/sdk-react-nextjs'],
+  experimental: {
+    serverComponentsExternalPackages: ['isolated-vm'],
+  },
 };
 
 module.exports = nextConfig;