+++ examples/next-js-sdk-gen-2-experimental-app-directory/src/app/[[...slug]]/page.tsx
@@ -10,6 +10,10 @@ interface MyPageProps {
 const apiKey = 'f1a790f8c3204b3b8c5c1795aeac4660';
 
 export default async function Page(props: MyPageProps) {
+  // NOTE: the import must be inside the Page component itself.
+  const { initializeNodeRuntime } = await import('@builder.io/sdk-react-nextjs/node/init');
+  initializeNodeRuntime();
+
   const urlPath = '/' + (props.params?.slug?.join('/') || '');
 
   const content = await fetchOneEntry({