+++ packages/sdks-tests/src/e2e-tests/symbols.spec.ts
@@ -60,8 +60,7 @@ const testSymbols = async (page: Page) => {
 };
 
 test.describe('Symbols', () => {
-  test('render correctly', async ({ page, sdk }) => {
-    test.fail(excludeTestFor({ angular: true }, sdk), 'Angular Gen2 SDK not implemented.');
+  test('render correctly', async ({ page }) => {
     await page.goto('/symbols');
 
     await testSymbols(page);