+++ packages/sdks/e2e/angular-ssr/src/app/app.component.ts
@@ -1,11 +1,19 @@
+import { isPlatformServer } from '@angular/common';
 // fails because type imports cannot be injected
 // eslint-disable-next-line @typescript-eslint/consistent-type-imports
-import { ChangeDetectorRef, Component } from '@angular/core';
+import {
+  ChangeDetectorRef,
+  Component,
+  Inject,
+  Optional,
+  PLATFORM_ID,
+} from '@angular/core';
 import {
   _processContentResult,
   fetchOneEntry,
   getBuilderSearchParams,
 } from '@builder.io/sdk-angular';
+import { REQUEST } from '@nguniversal/express-engine/tokens';
 import { getProps } from '@sdk/tests';
 import { HelloComponent } from './hello.component';
 
@@ -48,31 +56,36 @@ export class AppComponent {
     },
   ];
 
-  constructor(private cdr: ChangeDetectorRef) {}
-
-  async ngOnInit() {
-    const urlPath = window.location.pathname || '';
-
-    const builderProps = await getProps({
+  constructor(
+    private cdr: ChangeDetectorRef,
+    @Inject(PLATFORM_ID) private platformId: any,
+    @Optional() @Inject(REQUEST) private req: any
+  ) {
+    const urlPath = isPlatformServer(this.platformId)
+      ? this.req.path || ''
+      : window.location.pathname;
+    const searchParams = isPlatformServer(this.platformId)
+      ? new URLSearchParams(this.req.url.split('?')[1])
+      : new URLSearchParams(window.location.search);
+    getProps({
       pathname: urlPath,
       _processContentResult,
-      options: getBuilderSearchParams(
-        new URLSearchParams(window.location.search)
-      ),
+      options: searchParams ? getBuilderSearchParams(searchParams) : {},
       fetchOneEntry,
+    }).then((builderProps) => {
+      if (!builderProps) {
+        return;
+      }
+      this.content = builderProps.content;
+      this.canTrack = builderProps.canTrack;
+      this.trustedHosts = builderProps.trustedHosts;
+      this.apiKey = builderProps.apiKey;
+      this.model = builderProps.model;
+      this.apiVersion = builderProps.apiVersion;
     });
+  }
 
-    if (!builderProps) {
-      return;
-    }
-
-    this.content = builderProps.content;
-    this.canTrack = builderProps.canTrack;
-    this.trustedHosts = builderProps.trustedHosts;
-    this.apiKey = builderProps.apiKey;
-    this.model = builderProps.model;
-    this.apiVersion = builderProps.apiVersion;
-
+  async ngOnInit() {
     this.cdr.detectChanges();
   }
 }