+++ packages/sdks/e2e/nextjs-app-dir/src/components/MyTextBox/CatFacts.tsx
@@ -1,3 +1,5 @@
+'use server';
+
 export interface TextProps {
   text: string;
 }