+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @builder.io/sdk
 
+## 2.2.5
+
+### Patch Changes
+
+- a5b8810: Fix: handle parsing of stringified numeric values in `userAttributes`
+
 ## 2.2.4
 
 ### Patch Changes