+++ packages/sdks/output/angular/CHANGELOG.md
@@ -1,5 +1,73 @@
 # @builder.io/sdk-angular
 
+## 0.1.1
+
+### Patch Changes
+
+- 22de13c: Fix: add missing `override` component config
+
+## 0.1.0
+
+### Minor Changes
+
+- 3594120: Feature: add `shouldReceiveBuilderProps` config to Registered Components, with the following defaults:
+
+  ```ts
+  shouldReceiveBuilderProps: {
+      builderBlock: true,
+      builderContext: true,
+      builderComponents: false,
+      builderLinkComponent: false,
+    },
+  ```
+
+  To configure a component to receive only certain Builder props, override the `shouldReceiveBuilderProps` config:
+
+  Example:
+
+  ```ts
+  export const componentInfo = {
+    name: "Text",
+
+    shouldReceiveBuilderProps: {
+      builderBlock: true,
+      builderContext: false,
+      builderComponents: true,
+      builderLinkComponent: false,
+    },
+
+    inputs: [
+      {
+        name: "text",
+        type: "html",
+        required: true,
+        autoFocus: true,
+        bubble: true,
+        defaultValue: "Enter some text...",
+      },
+    ],
+  };
+  ```
+
+### Patch Changes
+
+- 27c2175: Feat: add multi bundle support for angular sdk (node and browser)
+- 5fb20a8: Fix: SSR for ab-tests and Symbols
+
+## 0.0.10
+
+### Patch Changes
+
+- 6187c39: Fix: `required` option for TextArea and Select blocks
+- 6187c39: Feat: Add support for TextArea block
+
+## 0.0.9
+
+### Patch Changes
+
+- bb4a5fd: Feature: add `webp` support for Image block file uploads.
+- 1f62b28: Fix: Remove `iframely` API key from Embed block logic.
+
 ## 0.0.8
 
 ### Patch Changes